+++ lib/src/models/fractal/attribute_object_enum.dart
@@ -38,4 +38,6 @@ enum AttributeObject {
   databaseHost,
   nest,
   egg,
+  @JsonValue('ssh_key')
+  sshKey,
 }