+++ lib/src/websocket/websocket_cubit.dart
@@ -42,8 +42,15 @@ class ServerWebsocketCubit extends Cubit<WebsocketState> with IWebsocketCubit {
   ServerWebsocketCubit({
     required this.client,
     required this.serverId,
+    bool autoInitialize = true,
   }) : super(const WebsocketState.initial()) {
-    init();
+    if (autoInitialize) init();
+  }
+
+  @override
+  void emit(state) {
+    if (isClosed) return;
+    super.emit(state);
   }
 
   @experimental
@@ -52,15 +59,15 @@ class ServerWebsocketCubit extends Cubit<WebsocketState> with IWebsocketCubit {
 
   @override
   close() async {
+    await _socket.close(WebSocketStatus.normalClosure);
     // close the specialized streams
     await listeners.closeAllListeners();
     // close the underlying websocket
-    await _socket.close();
     super.close();
   }
 
   /// Creates the listener that makes the individual streams work
-  void init() async {
+  Future<void> init() async {
     if (_isInitialized) return;
     await _connect();
     _configureListeners();