+++ CHANGELOG.md
@@ -1,4 +1,21 @@
 # CHANGELOG
+## 1.4.1
+ - added `lockBackup`
+ - updated server websocket extentions
+ - updated dependencies
+## 1.4.0
+ - bump release
+ - fixed an issue with the websocket cubit trying to emit new states after its closed
+
+## 1.4.0-dev.21
+ - renames `uploadFile` to `getFileUploadUrl`
+ - adds `ServerStateConflictException` error code and specifies an `UNKNOWN` when not in list
+
+## 1.4.0-dev.20
+ - Fix `CreateSshKey` model
+## 1.4.0-dev.19
+ - added `pullFile`, `createSshKey`, `listSshKeys`, and `chmodFile`
+
 ## 1.4.0-dev.18
  - Updated models so that `listdockerImages` actually works